@@ -1,14 +1,14 @@
+import 'package:bmi_calculator/components/bottom_button.dart';
+import 'package:bmi_calculator/components/reusable_card.dart';
 import 'package:bmi_calculator/constants.dart';
 import 'package:flutter/material.dart';
-import 'package:bmi_calculator/components/reusable_card.dart';
-import 'package:bmi_calculator/components/bottom_button.dart';
 import 'package:vibration/vibration.dart';
 
 class ResultsPage extends StatelessWidget {
   ResultsPage(
-      {@required this.bmiResult,
-      @required this.resultText,
-      @required this.interpretation});
+      {required this.bmiResult,
+      required this.resultText,
+      required this.interpretation});
   final String bmiResult;
   final String resultText;
   final String interpretation;
