@@ -1,13 +1,14 @@
 package co.appbrewery.bmicalculator;
 
-import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+// import android.os.Bundle; // No longer needed
+import io.flutter.embedding.android.FlutterActivity;
+// import io.flutter.plugins.GeneratedPluginRegistrant; // No longer needed
 
 public class MainActivity extends FlutterActivity {
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
-  }
+  // @Override
+  // protected void onCreate(Bundle savedInstanceState) {
+  //   super.onCreate(savedInstanceState);
+  //   GeneratedPluginRegistrant.registerWith(this); // Removed
+  // }
+  // No need to override onCreate if it's empty
 }
