@@ -4,7 +4,7 @@ import '../constants.dart';
 
 
 class IconContent extends StatelessWidget {
-  IconContent({this.icon, this.label});
+  IconContent({required this.icon, required this.label,});
   final IconData icon;
   final String label;
 
